pg_amcheck: Keep trying to fix the tests.
Commit 1d9152072203ffcbc77894e69720ce4c87a3105e managed to remove
one of the two places where we were checking for a "no such user"
error while leaving the other one right next to it. So remove that
too. In fact, remove the entire test, because the whole point of
this test was to see which message we got on a failure.